Upgrade Version Maven -> 3.9.9, Java -> 17

+++ identity_service/src/main/java/org/huyhieu/dto/data/UserDto.java
@@ -19,7 +19,7 @@ import java.util.Set;
 @Setter
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class UserDto {
-    long id;
+    Integer id;
     String username;
     String firstName;
     String lastName;
